+++ plugin/module-mapper.ts
@@ -1,4 +1,4 @@
-import { ModuleImport, ModuleMeta, ModulePart, ModuleLengths, ModuleUID } from "../types/types";
+import { ModuleImport, ModuleMeta, ModulePart, ModuleLengths, ModuleUID } from "../shared/types";
 import { getUid } from "./uid";
 
 const nanoid = getUid("1234567890abcdef", 4);
@@ -23,6 +23,9 @@ export class ModuleMapper {
   constructor(private projectRoot: string | RegExp) {}
 
   trimProjectRootId(moduleId: string): string {
+    if (typeof this.projectRoot === "string" && moduleId.startsWith(this.projectRoot)) {
+      return moduleId.slice(this.projectRoot.length);
+    }
     return moduleId.replace(this.projectRoot, "");
   }
 
@@ -66,11 +69,11 @@ export class ModuleMapper {
     if (uid in this.nodeParts) {
       throw new Error(
         `Override module: bundle id ${bundleId}, module id ${moduleId}, value ${JSON.stringify(
-          value
-        )}, existing value: ${JSON.stringify(this.nodeParts[uid])}`
+          value,
+        )}, existing value: ${JSON.stringify(this.nodeParts[uid])}`,
       );
     }
-    this.nodeParts[uid] = { ...value, mainUid: this.getModuleUid(moduleId) };
+    this.nodeParts[uid] = { ...value, metaUid: this.getModuleUid(moduleId) };
     return uid;
   }