+++ src/sunburst/node.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-unknown-property */
 import { HierarchyRectangularNode } from "d3-hierarchy";
 import { FunctionalComponent } from "preact";
-import { ModuleTree, ModuleTreeLeaf } from "../../types/types";
+import { ModuleTree, ModuleTreeLeaf } from "../../shared/types";
 import color from "../color";
 
 type NodeEventHandler = (event: HierarchyRectangularNode<ModuleTree | ModuleTreeLeaf>) => void;
@@ -25,6 +25,7 @@ export const Node: FunctionalComponent<NodeProps> = ({
 }) => {
   return (
     <path
+      className="node"
       d={path}
       fill-rule="evenodd"
       stroke="#fff"